requirements.txt
# Dependencies
feedparser>=6.0.0
yt-dlp>=2024.0.0
openai>=1.0.0
jinja2>=3.1.0
httpx>=0.25.0
pyyaml>=6.0
python-dateutil>=2.8.0